\name{ChangeLog}
\alias{ChangeLog}
\title{
  Log of Changes by \acronym{REFSfs} Version
}
\description{
Logs the significant changes that take place in each build.  Note that not all builds will be referenced here as many are generated by intermediate checkins that produce no meaningful change in features or functionality.  See the Git log for a full revision history for the code.

 	\describe{
		\item{Build <dont have a build number system yet>}{
                	\itemize{
                        	\item Created R package directories plugging into java mvn packaging, not yet building R package
                        	\item factored out major java packages to maven modules for cleaner feel
                	}
        	}
	}
}
